fix(subscriptions): address Phase 3 review findings
Six review findings against the Phase 3 PR, all verified against the code
and corrected:

1. **`_resolve_target_status` no longer un-terminalizes a CANCELLED row.**
   A late `customer.subscription.deleted` event that still carries
   `pause_collection` would have overwritten the local CANCELLED row to
   PAUSED (`cancelled_at` never set, since the assignment was gated on
   `target_status == CANCELLED`). PAUSED is not a TERMINAL_STATUS, so this
   re-armed the `one_active_subscription_per_user_org` partial unique
   index and would have blocked the user from resubscribing. Fix: map
   the Stripe `status` first and let CANCELLED/EXPIRED win over
   `pause_collection`.

2. **`_apply_stripe_price_swap` short-circuits on terminal rows.** Phase
   1's contract is that terminal subscriptions are frozen. The webhook
   sync now leaves `plan` / `pending_plan` / `stripe_schedule_id` alone
   when the local row is already terminal.

3. **ONLINE pause/resume refuse rows without `stripe_subscription_id`.**
   Previously the `payment_method == ONLINE AND stripe_subscription_id`
   guard let a Stripe-flavored row without a Stripe link silently fall
   through to the OFFLINE-only local pause path. Now the service raises
   400 in that state so we never mark a row PAUSED locally while Stripe
   keeps generating invoices.

4. **`_classify_plan_change` normalizes price by period.** Comparing
   `new_plan.price > current_plan.price` directly misclassified
   Monthly→Annual swaps (yearly headline is bigger even when per-month
   cost drops). The helper now divides by `_PERIOD_UNIT_MONTHS *
   period_count` so a 10/mo → 100/yr switch becomes a scheduled
   downgrade (8.33/mo effective) instead of an immediate prorated
   upgrade.

5. **`create_billing_portal_session` requires an existing CustomerProfile.**
   The endpoint used to lazy-create the Stripe Customer on first use,
   so any authenticated user could iterate org IDs and spam junk
   Customer rows on every Connect account. Now we look the profile up
   and 404 if it's missing — only users who have actually subscribed
   can get a portal session.

6. **`BillingPortalRequestSchema.return_url` is validated as `HttpUrl`.**
   `javascript:` / `data:` / malformed URLs are rejected at the schema
   layer (422). Controller casts the validated value back to `str` for
   the Stripe API.

Tests
- `test_subscription_service.py`: ONLINE pause/resume refusal without a
  Stripe id.
- `test_subscription_stripe_service.py` (parent): no new tests in the
  parent file. Existing Phase 3 sections were moved to a new sibling
  file to keep both under the 1000-line file-length limit.
- `test_subscription_stripe_service_phase3.py` (new): all Phase 3
  service tests, including the new coverage for the fixes —
  terminal-status-wins-over-pause-collection, price-swap-skipped-on-
  terminal, monthly↔annual cross-cadence classification (downgrade
  cheaper-per-month / upgrade pricier-per-month), and billing-portal
  refusal when no CustomerProfile exists.
- `test_me_subscriptions.py`: billing-portal endpoint now requires a
  seeded CustomerProfile; new `test_rejects_invalid_return_url` covers
  `HttpUrl` validation; the happy-path assertion was loosened to
  `.startswith(...)` since `HttpUrl` normalizes the URL on the wire.

`make check` is green (format, lint, mypy strict, migration-check,
i18n, file-length).
