+++ src/events/tests/test_service/test_subscription_service.py
@@ -346,6 +346,56 @@ class TestLifecycle:
         out = subscription_service.cancel_subscription(sub, immediate=True)
         assert out.status == MembershipSubscription.SubscriptionStatus.CANCELLED
 
+    def test_pause_online_without_stripe_id_is_refused(
+        self,
+        tier: MembershipTier,
+        subscriber: RevelUser,
+    ) -> None:
+        """An ONLINE row that hasn't been linked to a Stripe subscription must not pause locally."""
+        online_plan = MembershipSubscriptionPlan.objects.create(
+            tier=tier,
+            name="Online",
+            price=Decimal("10.00"),
+            currency="EUR",
+            period_unit="month",
+            period_count=1,
+            payment_method=MembershipSubscriptionPlan.PaymentMethod.ONLINE,
+        )
+        # stripe_subscription_id intentionally empty
+        sub = MembershipSubscription.objects.create(
+            user=subscriber,
+            plan=online_plan,
+            organization=online_plan.tier.organization,
+            status=MembershipSubscription.SubscriptionStatus.ACTIVE,
+        )
+        with pytest.raises(HttpError) as exc:
+            subscription_service.pause_subscription(sub)
+        assert exc.value.status_code == 400
+
+    def test_resume_online_without_stripe_id_is_refused(
+        self,
+        tier: MembershipTier,
+        subscriber: RevelUser,
+    ) -> None:
+        online_plan = MembershipSubscriptionPlan.objects.create(
+            tier=tier,
+            name="Online R",
+            price=Decimal("10.00"),
+            currency="EUR",
+            period_unit="month",
+            period_count=1,
+            payment_method=MembershipSubscriptionPlan.PaymentMethod.ONLINE,
+        )
+        sub = MembershipSubscription.objects.create(
+            user=subscriber,
+            plan=online_plan,
+            organization=online_plan.tier.organization,
+            status=MembershipSubscription.SubscriptionStatus.PAUSED,
+        )
+        with pytest.raises(HttpError) as exc:
+            subscription_service.resume_subscription(sub)
+        assert exc.value.status_code == 400
+
 
 # ---- refund_payment ----------------------------------------------------------
