+++ src/events/tests/test_service/test_subscription_stripe_service.py
@@ -572,394 +572,3 @@ class TestRecordStripePaymentFromInvoice:
             "currency": "eur",
         }
         assert subscription_stripe_service.record_stripe_payment_from_invoice(invoice, succeeded=False) is None
-
-
-# ---- change_online_plan (Phase 3) -------------------------------------------
-
-
-def _make_online_subscription(
-    plan: MembershipSubscriptionPlan,
-    user: RevelUser,
-    *,
-    stripe_id: str = "sub_for_change",
-) -> MembershipSubscription:
-    """Create an ACTIVE ONLINE subscription with a Stripe link, ready for plan-change tests."""
-    return MembershipSubscription.objects.create(
-        user=user,
-        plan=plan,
-        organization=plan.tier.organization,
-        status=MembershipSubscription.SubscriptionStatus.ACTIVE,
-        stripe_subscription_id=stripe_id,
-        current_period_end=timezone.now() + timedelta(days=30),
-    )
-
-
-class TestChangeOnlinePlan:
-    @pytest.fixture
-    def pricier_plan(self, tier: MembershipTier) -> MembershipSubscriptionPlan:
-        """A second ONLINE plan on the same tier, twice as expensive."""
-        return MembershipSubscriptionPlan.objects.create(
-            tier=tier,
-            name="Yearly Online",
-            price=Decimal("20.00"),
-            currency="EUR",
-            period_unit="year",
-            period_count=1,
-            payment_method=MembershipSubscriptionPlan.PaymentMethod.ONLINE,
-            stripe_product_id="prod_year",
-            stripe_price_id="price_year",
-        )
-
-    @pytest.fixture
-    def cheaper_plan(self, tier: MembershipTier) -> MembershipSubscriptionPlan:
-        """A second ONLINE plan on the same tier, half the price (downgrade target)."""
-        return MembershipSubscriptionPlan.objects.create(
-            tier=tier,
-            name="Lite Online",
-            price=Decimal("5.00"),
-            currency="EUR",
-            period_unit="month",
-            period_count=1,
-            payment_method=MembershipSubscriptionPlan.PaymentMethod.ONLINE,
-            stripe_product_id="prod_lite",
-            stripe_price_id="price_lite",
-        )
-
-    @mock.patch("events.service.subscription_stripe_service.stripe.Subscription.modify")
-    @mock.patch("events.service.subscription_stripe_service.stripe.Subscription.retrieve")
-    def test_upgrade_modifies_stripe_with_proration(
-        self,
-        mock_retrieve: mock.Mock,
-        mock_modify: mock.Mock,
-        online_plan: MembershipSubscriptionPlan,
-        pricier_plan: MembershipSubscriptionPlan,
-        subscriber: RevelUser,
-    ) -> None:
-        sub = _make_online_subscription(online_plan, subscriber)
-        mock_retrieve.return_value = {"items": {"data": [{"id": "si_test"}]}}
-
-        result = subscription_stripe_service.change_online_plan(sub, pricier_plan)
-
-        mock_modify.assert_called_once()
-        kwargs = mock_modify.call_args.kwargs
-        assert kwargs["items"] == [{"id": "si_test", "price": "price_year"}]
-        assert kwargs["proration_behavior"] == "create_prorations"
-        assert kwargs["payment_behavior"] == "allow_incomplete"
-        assert kwargs["stripe_account"] == "acct_test_org"
-        result.refresh_from_db()
-        assert result.plan_id == pricier_plan.pk
-        assert result.pending_plan_id is None
-
-    @mock.patch("events.service.subscription_stripe_service.stripe.SubscriptionSchedule.modify")
-    @mock.patch("events.service.subscription_stripe_service.stripe.SubscriptionSchedule.create")
-    def test_downgrade_creates_schedule_and_sets_pending_plan(
-        self,
-        mock_schedule_create: mock.Mock,
-        mock_schedule_modify: mock.Mock,
-        online_plan: MembershipSubscriptionPlan,
-        cheaper_plan: MembershipSubscriptionPlan,
-        subscriber: RevelUser,
-    ) -> None:
-        sub = _make_online_subscription(online_plan, subscriber)
-        schedule_mock = mock.MagicMock(id="sub_sched_1")
-        schedule_mock.get.return_value = [{"items": [{"price": "price_test"}], "start_date": 100, "end_date": 200}]
-        mock_schedule_create.return_value = schedule_mock
-
-        result = subscription_stripe_service.change_online_plan(sub, cheaper_plan)
-
-        mock_schedule_create.assert_called_once()
-        mock_schedule_modify.assert_called_once()
-        phases = mock_schedule_modify.call_args.kwargs["phases"]
-        assert phases[0]["items"][0]["price"] == "price_test"
-        assert phases[1]["items"][0]["price"] == "price_lite"
-        assert mock_schedule_modify.call_args.kwargs["end_behavior"] == "release"
-        result.refresh_from_db()
-        # Downgrade leaves current plan in place — the pending plan is queued.
-        assert result.plan_id == online_plan.pk
-        assert result.pending_plan_id == cheaper_plan.pk
-        assert result.stripe_schedule_id == "sub_sched_1"
-
-    def test_refuses_same_plan(
-        self,
-        online_plan: MembershipSubscriptionPlan,
-        subscriber: RevelUser,
-    ) -> None:
-        sub = _make_online_subscription(online_plan, subscriber)
-        with pytest.raises(HttpError) as exc:
-            subscription_stripe_service.change_online_plan(sub, online_plan)
-        assert exc.value.status_code == 400
-
-    def test_refuses_when_pending_plan_already_set(
-        self,
-        online_plan: MembershipSubscriptionPlan,
-        cheaper_plan: MembershipSubscriptionPlan,
-        subscriber: RevelUser,
-    ) -> None:
-        sub = _make_online_subscription(online_plan, subscriber)
-        sub.pending_plan = cheaper_plan
-        sub.save(update_fields=["pending_plan"])
-        with pytest.raises(HttpError) as exc:
-            subscription_stripe_service.change_online_plan(sub, cheaper_plan)
-        assert exc.value.status_code == 400
-
-    def test_refuses_when_paused(
-        self,
-        online_plan: MembershipSubscriptionPlan,
-        cheaper_plan: MembershipSubscriptionPlan,
-        subscriber: RevelUser,
-    ) -> None:
-        sub = _make_online_subscription(online_plan, subscriber)
-        sub.status = MembershipSubscription.SubscriptionStatus.PAUSED
-        sub.save(update_fields=["status"])
-        with pytest.raises(HttpError) as exc:
-            subscription_stripe_service.change_online_plan(sub, cheaper_plan)
-        assert exc.value.status_code == 400
-
-    def test_refuses_when_cancel_at_period_end(
-        self,
-        online_plan: MembershipSubscriptionPlan,
-        cheaper_plan: MembershipSubscriptionPlan,
-        subscriber: RevelUser,
-    ) -> None:
-        sub = _make_online_subscription(online_plan, subscriber)
-        sub.cancel_at_period_end = True
-        sub.save(update_fields=["cancel_at_period_end"])
-        with pytest.raises(HttpError) as exc:
-            subscription_stripe_service.change_online_plan(sub, cheaper_plan)
-        assert exc.value.status_code == 400
-
-
-# ---- pause_online_subscription / resume_online_subscription -----------------
-
-
-class TestPauseResumeOnlineSubscription:
-    @pytest.fixture
-    def online_subscription(
-        self,
-        online_plan: MembershipSubscriptionPlan,
-        subscriber: RevelUser,
-    ) -> MembershipSubscription:
-        return _make_online_subscription(online_plan, subscriber, stripe_id="sub_pause_test")
-
-    @mock.patch("events.service.subscription_stripe_service.stripe.Subscription.modify")
-    def test_pause_calls_stripe_with_void_behavior(
-        self,
-        mock_modify: mock.Mock,
-        online_subscription: MembershipSubscription,
-    ) -> None:
-        result = subscription_stripe_service.pause_online_subscription(online_subscription)
-        mock_modify.assert_called_once_with(
-            "sub_pause_test",
-            pause_collection={"behavior": "void"},
-            stripe_account="acct_test_org",
-        )
-        assert result.status == MembershipSubscription.SubscriptionStatus.PAUSED
-
-    @mock.patch("events.service.subscription_stripe_service.stripe.Subscription.modify")
-    def test_pause_is_idempotent_when_already_paused(
-        self,
-        mock_modify: mock.Mock,
-        online_subscription: MembershipSubscription,
-    ) -> None:
-        online_subscription.status = MembershipSubscription.SubscriptionStatus.PAUSED
-        online_subscription.save(update_fields=["status"])
-        result = subscription_stripe_service.pause_online_subscription(online_subscription)
-        mock_modify.assert_not_called()
-        assert result.status == MembershipSubscription.SubscriptionStatus.PAUSED
-
-    def test_pause_refuses_offline(
-        self,
-        offline_plan: MembershipSubscriptionPlan,
-        subscriber: RevelUser,
-    ) -> None:
-        sub = MembershipSubscription.objects.create(
-            user=subscriber,
-            plan=offline_plan,
-            organization=offline_plan.tier.organization,
-            status=MembershipSubscription.SubscriptionStatus.ACTIVE,
-        )
-        with pytest.raises(HttpError) as exc:
-            subscription_stripe_service.pause_online_subscription(sub)
-        assert exc.value.status_code == 400
-
-    @mock.patch("events.service.subscription_stripe_service.stripe.Subscription.modify")
-    def test_pause_stripe_error_propagates_502(
-        self,
-        mock_modify: mock.Mock,
-        online_subscription: MembershipSubscription,
-    ) -> None:
-        mock_modify.side_effect = stripe.error.APIConnectionError("boom")
-        with pytest.raises(HttpError) as exc:
-            subscription_stripe_service.pause_online_subscription(online_subscription)
-        assert exc.value.status_code == 502
-        online_subscription.refresh_from_db()
-        # Local state stays untouched when Stripe call fails.
-        assert online_subscription.status == MembershipSubscription.SubscriptionStatus.ACTIVE
-
-    @mock.patch("events.service.subscription_stripe_service.stripe.Subscription.modify")
-    def test_resume_clears_pause_collection(
-        self,
-        mock_modify: mock.Mock,
-        online_subscription: MembershipSubscription,
-    ) -> None:
-        online_subscription.status = MembershipSubscription.SubscriptionStatus.PAUSED
-        online_subscription.save(update_fields=["status"])
-        result = subscription_stripe_service.resume_online_subscription(online_subscription)
-        mock_modify.assert_called_once_with(
-            "sub_pause_test",
-            pause_collection="",
-            stripe_account="acct_test_org",
-        )
-        assert result.status == MembershipSubscription.SubscriptionStatus.ACTIVE
-
-    def test_resume_refuses_non_paused(
-        self,
-        online_subscription: MembershipSubscription,
-    ) -> None:
-        with pytest.raises(HttpError) as exc:
-            subscription_stripe_service.resume_online_subscription(online_subscription)
-        assert exc.value.status_code == 400
-
-
-# ---- create_billing_portal_session ------------------------------------------
-
-
-class TestCreateBillingPortalSession:
-    @mock.patch("events.service.subscription_stripe_service.stripe.billing_portal.Session.create")
-    @mock.patch("events.service.subscription_stripe_service.stripe.Customer.create")
-    def test_returns_session_url(
-        self,
-        mock_customer: mock.Mock,
-        mock_portal: mock.Mock,
-        stripe_org: Organization,
-        subscriber: RevelUser,
-    ) -> None:
-        mock_customer.return_value = mock.MagicMock(id="cus_for_portal")
-        mock_portal.return_value = mock.MagicMock(url="https://stripe.example/portal/abc")
-
-        url = subscription_stripe_service.create_billing_portal_session(
-            subscriber, stripe_org, return_url="https://app.example/account"
-        )
-
-        assert url == "https://stripe.example/portal/abc"
-        kwargs = mock_portal.call_args.kwargs
-        assert kwargs["customer"] == "cus_for_portal"
-        assert kwargs["return_url"] == "https://app.example/account"
-        assert kwargs["stripe_account"] == "acct_test_org"
-
-    def test_refuses_non_connected_org(
-        self,
-        organization: Organization,
-        subscriber: RevelUser,
-    ) -> None:
-        with pytest.raises(HttpError) as exc:
-            subscription_stripe_service.create_billing_portal_session(
-                subscriber, organization, return_url="https://app.example/account"
-            )
-        assert exc.value.status_code == 400
-
-    @mock.patch("events.service.subscription_stripe_service.stripe.billing_portal.Session.create")
-    @mock.patch("events.service.subscription_stripe_service.stripe.Customer.create")
-    def test_stripe_failure_propagates_502(
-        self,
-        mock_customer: mock.Mock,
-        mock_portal: mock.Mock,
-        stripe_org: Organization,
-        subscriber: RevelUser,
-    ) -> None:
-        mock_customer.return_value = mock.MagicMock(id="cus_portal")
-        mock_portal.side_effect = stripe.error.APIConnectionError("boom")
-        with pytest.raises(HttpError) as exc:
-            subscription_stripe_service.create_billing_portal_session(
-                subscriber, stripe_org, return_url="https://app.example/account"
-            )
-        assert exc.value.status_code == 502
-
-
-# ---- sync_subscription_from_stripe — Phase 3 additions ----------------------
-
-
-class TestSyncPauseCollectionAndPlanSwap:
-    @pytest.fixture
-    def online_subscription(
-        self,
-        online_plan: MembershipSubscriptionPlan,
-        subscriber: RevelUser,
-    ) -> MembershipSubscription:
-        return _make_online_subscription(online_plan, subscriber, stripe_id="sub_phase3_sync")
-
-    def test_pause_collection_forces_paused_even_when_status_is_active(
-        self,
-        online_subscription: MembershipSubscription,
-    ) -> None:
-        payload = {
-            "id": "sub_phase3_sync",
-            "status": "active",
-            "pause_collection": {"behavior": "void"},
-            "cancel_at_period_end": False,
-        }
-        result = subscription_stripe_service.sync_subscription_from_stripe(payload)
-        assert result is not None
-        online_subscription.refresh_from_db()
-        assert online_subscription.status == MembershipSubscription.SubscriptionStatus.PAUSED
-
-    def test_resume_clears_pause_and_uses_mapped_status(
-        self,
-        online_subscription: MembershipSubscription,
-    ) -> None:
-        online_subscription.status = MembershipSubscription.SubscriptionStatus.PAUSED
-        online_subscription.save(update_fields=["status"])
-        payload = {
-            "id": "sub_phase3_sync",
-            "status": "active",
-            "pause_collection": None,
-        }
-        subscription_stripe_service.sync_subscription_from_stripe(payload)
-        online_subscription.refresh_from_db()
-        assert online_subscription.status == MembershipSubscription.SubscriptionStatus.ACTIVE
-
-    def test_price_swap_repoints_plan_and_clears_pending(
-        self,
-        online_plan: MembershipSubscriptionPlan,
-        online_subscription: MembershipSubscription,
-        tier: MembershipTier,
-    ) -> None:
-        new_plan = MembershipSubscriptionPlan.objects.create(
-            tier=tier,
-            name="Switched",
-            price=Decimal("7.00"),
-            currency="EUR",
-            period_unit="month",
-            period_count=1,
-            payment_method=MembershipSubscriptionPlan.PaymentMethod.ONLINE,
-            stripe_product_id="prod_sw",
-            stripe_price_id="price_switched",
-        )
-        online_subscription.pending_plan = new_plan
-        online_subscription.stripe_schedule_id = "sub_sched_phase"
-        online_subscription.save(update_fields=["pending_plan", "stripe_schedule_id"])
-
-        payload = {
-            "id": "sub_phase3_sync",
-            "status": "active",
-            "items": {"data": [{"price": {"id": "price_switched"}}]},
-        }
-        subscription_stripe_service.sync_subscription_from_stripe(payload)
-        online_subscription.refresh_from_db()
-        assert online_subscription.plan_id == new_plan.pk
-        assert online_subscription.pending_plan_id is None
-        assert online_subscription.stripe_schedule_id == ""
-
-    def test_price_swap_to_unknown_price_is_ignored(
-        self,
-        online_subscription: MembershipSubscription,
-    ) -> None:
-        payload = {
-            "id": "sub_phase3_sync",
-            "status": "active",
-            "items": {"data": [{"price": {"id": "price_unknown_to_us"}}]},
-        }
-        subscription_stripe_service.sync_subscription_from_stripe(payload)
-        online_subscription.refresh_from_db()
-        # Plan stays unchanged; we don't blindly orphan the subscription.
-        assert online_subscription.plan_id == online_subscription.plan_id
